Change to a subdomain without underscore
Let's Encrypt does not support subdomains with underscores as they
are not valid hostnames according to the RFCs. Although _pine is a
short and concise subdomain, it would make it cumbersome to obtain
an SSL certificate for it, which is required for Pine. So for this
reason the subdomain has been changed to not include any underscores
and is therefore longer to guarantee its uniqueness.

<https://community.letsencrypt.org/t/underscore-in-subdomain-fails/31431>